Add new temperature processing technique
Previously, temperature averaging was simply averaging. However, for
many polymers, the cloud point (when it condenses) is different than the
clear point (when it dissolves). As such, now we calculate
if the temperature is in a cooling stage, heating stage, or ramping
stage and encode it as a one-hot encoding. This can help prevent some
noise in the transmission percentage that is caused by this difference.

__Implementation__
1. Calculate the time step that occurs between each temperature index.
2. At each point, take avereage 30 seconds worth of prior and post temperatures. If
prior temperature is less than current, and post temperature is greater,
it's in heating state. Opposite is cooling state, else, it is hold
state.
3. Take average of specific temperatures based on the state they are in.
4. Alter database encoding to account for this change in variable.