Fix bug in temperature program loading
Assumption I made about when loading stage stopped and experiment stage
started was incorrect. Initially, I thought after "Stir rate" occurred,
change in temperature program state happened, but in reality this was
not the case.

Now, I just switch from "sammple_load" to "experiment" after the first
temperature program step occurs that is part of the sample load phase.
If this is also an incorrect assumption, that is unfortunate and I can
come up with a method where we skip the first data points that are
associated with the "holding" ramp state instead.