Implement filtering of data
An issue that persists with the new averaging technique is that we see
well defined behavior in some regions but register large uncertainties
due to intrinsic noise. We may want to throw out some data points that
have high uncertainties due to solubility transition shifts or
unequilibrated states, but not high uncertainties here.

I implemented a Savitzky-Golay filter, which is commonly used in signal
processing and spectroscopy. I was expecting linear behavior with
temperature, so used a polynomial order of one, and chose the number of
data points based on 2 minutes worth of data.

Added filtered column to processed df indicating if the transmission is
after filtering data or not and added flag to database.